llama-swap.go: swap main entry point
